Add task to ease checking of version mappings
Previously, to check that all of a project's dependencies had Spring IO
version mappings it was necessary to run the springIoCheck tasks,
specifying a JDK 7 and/or JDK 8 home, or to run the dependencies task
on each module.

To make this process easier, this commit adds a task, that the
springioCheck task depends on, which causes the springioTestRuntime
configuration to be resolved thereby triggering a check of its
dependency version mapping.