Perform version checking and version mapping against separate configs
Previously the springioTestRuntime configuration was used both when
checking that all of a project's dependencies were part of the
platform and when modifying those dependencies to use the platform's
versions. This meant that test-only dependencies had to be included
in the Platform.

This commit updates the plugin to use separate configuration for
checking that a dependency is in the platform. By default the checking
will be performed against the runtime configuration that's provided by
the Java plugin. The task can be configured to use a different
configuration:

springioDependencyVersionMappingCheck {
	configuration = configurations.configurationToCheck
}

The configuration options for controlling whether or not unmapped
direct dependencies and unmapped transitive dependencies should cause
a failure have been moved from the Spring IO platform extension onto
this task. These options can be configured as follows:

springioDependencyVersionMappingCheck {
	failOnUnmappedDirectDependency = true
	failOnUnmappedTransitiveDependency = true
}

The Spring IO platform extension has been removed.