Remove @Input from Configuration properties
Gradle does not support properties that hold configuration instances
being annotated as inputs. Doing so may result in a failure as
Gradle tries to serialize the Configuration for caching. This commit
removes @Input from task properties that hold a Configuration. It
also removes @OutputFile from affected tasks. Without this change
Gradle would incorrectly assume that the task was up-to-date as the
contents of the output had not changed since the previous run.

Note: a downside of this change is that the task will now always be
run as Gradle no longer has the information required to determine
whether or not it's up-to-date.

Fixes #39