Update configuration instructions
Update the instructions for configuring the platform to recommend the
use of an external property for the version of the platform-versions
dependency and to only apply the plugin when that property is set.

This configuration approach means that the any version of a project can
be checked out and tested against any version of the Platform.